pyproject.toml
[project]
name = "bulk-calendar-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-dance[sqla]>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-flask>=1.3.0",
    "pytest-mock>=3.15.1",
]
